Fix build break:
IDEA tests do not pick up product changes until building them with gradle.py, I missed a testcase.
By running the tests, I noticed that GSON roundtrips numbers as Long, so minAPi needs to be typed as such.

Bug:

+++ src/main/java/com/android/tools/r8/dex/Marker.java
@@ -68,11 +68,11 @@ public class Marker {
   }
 
   @Nullable
-  public Integer getMinApi() {
-    return (Integer) content.get(MIN_API);
+  public Long getMinApi() {
+    return (Long) content.get(MIN_API);
   }
 
-  public Marker setMinApi(int minApi) {
+  public Marker setMinApi(long minApi) {
     internalPut(MIN_API, minApi);
     return this;
   }
